// apps/web/lib/api-types.generated.ts
export interface components {
    schemas: {
        /** LivenessResponse */
        LivenessResponse: {
            /**
             * Status
             * @constant
             */
            status: "alive";
        };
        /** ReadinessResponse */
        ReadinessResponse: {
            /**
             * Status
             * @enum {string}
             */
            status: "ready" | "degraded";
            /** App Env */
            app_env: string;
            /** Dependencies */
            dependencies: {
                [key: string]: string;
            };
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}